-Modified request for search store. -Added additional keys from api.

+++ Core/core/src/main/java/com/yilinker/core/model/buyer/Aggregation.java
@@ -7,23 +7,23 @@ import java.util.List;
  */
 public class Aggregation {
 
-    private int maxPrice,minPrice;
+    private double maxPrice,minPrice;
 
     private List<FilterGroup> attributes;
 
-    public int getMaxPrice() {
+    public double getMaxPrice() {
         return maxPrice;
     }
 
-    public void setMaxPrice(int maxPrice) {
+    public void setMaxPrice(double maxPrice) {
         this.maxPrice = maxPrice;
     }
 
-    public int getMinPrice() {
+    public double getMinPrice() {
         return minPrice;
     }
 
-    public void setMinPrice(int minPrice) {
+    public void setMinPrice(double minPrice) {
         this.minPrice = minPrice;
     }
